complex dims eval error bug fix
if the first complex dim in the aggr result evaluates to an empty
string, ngx_aggr_result_temp_alloc returns null (requested size is not
larger than allocated size, both are zero)
this makes ngx_aggr_complex_value fail.
the fix is to allocate the buffer if requested size is bigger than
allocated OR the current pointer is null.